refactor(UserChallenges): moving Item and Widget components to their own directories
