+++ src/modules/UserChallenges/Item/UserChallengeItem.styles.ts
