+++ src/modules/UserChallenges/Item/index.ts
@@ -1,0 +1,3 @@
+import UserChallengeItem from './UserChallengeItem'
+
+export default UserChallengeItem
