+++ src/modules/UserChallenges/Widget/UserChallengesWidget.container.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import { useSelector } from 'react-redux'
 
-import selector from './UserChallenges.selector'
-import UserChallengesWidget from './UserChallenges.widget'
+import selector from '../UserChallenges.selector'
+import UserChallengesWidget from './UserChallengesWidget'
 
 const UserChallengesWidgetContainer = () => {
   const { challenges } = useSelector(selector)
