+++ src/modules/UserChallenges/Widget/index.ts
@@ -1,0 +1,3 @@
+import UserChallengesWidgetContainer from './UserChallengesWidget.container'
+
+export default UserChallengesWidgetContainer
